Organizing augmentation tests into the subdirectories _3d and container and moving utility function (which got renamed from reproducability_test.py to utils.py) to testing/augmentation
