+++ tests/augmentation/_3d/test_center_crop_3d.py
