+++ tests/augmentation/_3d/test_random_equalize_3d.py
