+++ tests/augmentation/container/test_patch_sequential.py
@@ -20,7 +20,7 @@ import torch
 
 import kornia.augmentation as K
 
-from reproducibility_test import reproducibility_test
+from testing.augmentation.utils import reproducibility_test
 
 class TestPatchSequential:
     @pytest.mark.parametrize(
